# server_side/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Domentation vehicle tracking API",
        "version": "1.0.0",
        "title": "Vehicle tracking APIs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "anglebertsh@gmail.com",
            "name": "Ishimwe SHumbusho Anglebert"
        },
        "licence": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
...
    "paths": {
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register admin",
                "description": "Register admin",
                "operationId": "registerAdmin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Admin object",
                        "required": true,
                        "properties": {
                            "names": {
                                "type": "string",
                                "example": "anglebert"
                            },
                            "phoneNumber": {
                                "type": "string",
                                "example": "0788781096"
                            },
                            "email": {
                                "type": "string",
                                "example": "123@gmail.com"
                            },
                            "password": {
                                "type": "string"
                            },
                            "reEnterPassword": {
                                "type": "string",
                                "example": "string"
                            },
                            "nationalID": {
                                "type": "string",
                                "example": "1199999999999999"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/user/current": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the current User",
                "description": "Get the current user",
                "operationId": "getCurrentUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    }
}